Refactor MariaDB configuration and discovery modules
- Moved MariaDB configuration types and functions to a new `config` package.
- Updated installation and removal processes to use the new configuration structure.
- Replaced direct references to `mariadb` package with `discovery` package for installation discovery.
- Removed deprecated `discovery_wrapper` and `flags` files, consolidating functionality into the new structure.
- Enhanced validation logic for configuration inputs.
- Improved logging and error handling throughout the installation and removal processes.
