+++ internal/core/mariadb/install/service.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"sfDBTools/internal/logger"
-	"sfDBTools/utils/mariadb"
+	mariadb_config "sfDBTools/utils/mariadb/config"
 	defaultsetup "sfDBTools/utils/mariadb/defaultSetup"
 	"sfDBTools/utils/terminal"
 )
@@ -30,7 +30,7 @@ func startMariaDBService(deps *defaultsetup.Dependencies) error {
 }
 
 // verifyInstallation memverifikasi bahwa instalasi berhasil
-func verifyInstallation(cfg *mariadb.MariaDBInstallConfig, deps *defaultsetup.Dependencies) error {
+func verifyInstallation(cfg *mariadb_config.MariaDBInstallConfig, deps *defaultsetup.Dependencies) error {
 	lg, _ := logger.Get()
 
 	// Cek apakah service berjalan
@@ -52,7 +52,7 @@ func verifyInstallation(cfg *mariadb.MariaDBInstallConfig, deps *defaultsetup.De
 }
 
 // displaySuccessMessage menampilkan pesan sukses dan instruksi
-func displaySuccessMessage(cfg *mariadb.MariaDBInstallConfig) {
+func displaySuccessMessage(cfg *mariadb_config.MariaDBInstallConfig) {
 	terminal.SafePrintln("\n🎉 Instalasi MariaDB berhasil!")
 	terminal.SafePrintln(fmt.Sprintf("   Versi: %s", cfg.Version))
 	terminal.SafePrintln("\n📝 Langkah selanjutnya:")
