+++ utils/mariadb/discovery/data_socket.go
@@ -1,4 +1,4 @@
-package mariadb
+package discovery
 
 import (
 	"fmt"
