+++ utils/mariadb/discovery/service.go
@@ -1,4 +1,4 @@
-package mariadb
+package discovery
 
 import (
 	"fmt"
