+++ utils/mariadb/discovery/util.go
@@ -1,4 +1,4 @@
-package mariadb
+package discovery
 
 import "regexp"
